requirements.txt
adal==1.2.4
asgiref==3.2.8
azure-common==1.1.25
azure-core==1.6.0
azure-mgmt-storage==11.1.0
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
boto3==1.14.12
botocore==1.17.12
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.7
django-allauth==0.42.0
django-ckeditor==5.9.0
django-crispy-forms==1.9.1
django-js-asset==1.2.2
django-multiupload==0.5.2
django-realtime==1.1
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
idna==2.9
isodate==0.6.0
jmespath==0.10.0
msrest==0.6.17
msrestazure==0.6.3
oauthlib==3.1.0
Pillow==7.1.2
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
simplejson==3.17.0
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
